requirements.txt
torch
torchvision
torchaudio
transformers
sentence-transformers
numpy
pillow
einops
accelerate
qwen-vl-utils